processor/requirements.txt
kafka-python==2.0.2
pymongo==4.6.0
flask==3.0.0
werkzeug==3.0.1
python-snappy
prometheus-client==0.19.0
